reorganise services
