+++ components/products/mobile-chat-tab.tsx
@@ -6,7 +6,7 @@ import { useChatSession, useSendMessage } from '@/hooks/use-chat'
 import { Loader2, MessageSquareText, BookOpen, GraduationCap, HelpCircle } from 'lucide-react'
 import { Message, type MessageType } from '@/components/assistant/message'
 import { MessageInput } from '@/components/assistant/message-input'
-import type { GetSessionResult, ChatMessage } from '@/lib/chat.service'
+import type { GetSessionResult, ChatMessage } from '@/lib/service/chat.service'
 
 interface MobileChatTabProps {
   sessionId: string | null
